/ fixes bug id 2, inactivity accepts floating point
--HG--
extra : convert_revision : svn%3A05628a4a-d5d7-11dd-a2b5-e5f827957e07/trunk%40110

+++ log4tailer/Actions/InactivityAction.py
@@ -34,7 +34,7 @@ class InactivityAction:
     def triggerAction(self,message):
         if not message.getPlainMessage():
             ellapsedTime = self.timer.inactivityEllapsed()
-            if ellapsedTime > int(self.inactivityTime):
+            if ellapsedTime > float(self.inactivityTime):
                 self.acumulativeTime += ellapsedTime
                 # at this moment we will just print an emphasized alert in stdout
                 print self.logColors.backgroundemph+"Inactivity in the log for "+ str(self.acumulativeTime) + " seconds"+self.logColors.reset
